refactor(languages): Onboard French and Catalan to new module layout
